add comment delete functionality with user verification

+++ src/comments/views.py
@@ -1,6 +1,6 @@
 from django.contrib import messages
 from django.contrib.contenttypes.models import ContentType
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404, HttpResponse
 from django.shortcuts import render, get_object_or_404
 
 from .forms import CommentForm
@@ -8,7 +8,19 @@ from .models import Comment
 
 
 def comment_delete(request, pk):
-    obj = get_object_or_404(Comment, pk=pk)
+    try:
+        obj = Comment.objects.get(pk=pk)
+    except:
+        raise Http404
+
+    if obj.user != request.user:
+        # messages.success(request, "You do not have permission to view this.")
+        # raise Http404
+        response = HttpResponse("You do not have permission to do this.")
+        response.status_code = 403
+        return response
+
+
     if request.method == "POST":
         parent_obj_url = obj.content_object.get_absolute_url()
         obj.delete()
@@ -21,7 +33,14 @@ def comment_delete(request, pk):
 
 
 def comment_thread(request, pk):
-    obj = get_object_or_404(Comment, pk=pk)
+    try:
+        obj = Comment.objects.get(pk=pk)
+    except:
+        raise Http404
+
+    if not obj.is_parent:
+        obj = obj.parent
+
     content_object = obj.content_object # Post that the comment is on
     content_id = obj.content_object.id
 
@@ -45,7 +64,6 @@ def comment_thread(request, pk):
             parent_qs = Comment.objects.filter(id=parent_id)
             if parent_qs.exists() and parent_qs.count() == 1:
                 parent_obj = parent_qs.first()
-
 
         new_comment, created = Comment.objects.get_or_create(
                             user = request.user,
